Split the ServerManager.java Added auto restarting for servers
