+++ src/main/java/dev/gilbert/zacharia/managercraft/ManagercraftApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.scheduling.annotation.EnableAsync;
 
-@SpringBootApplication
 @EnableAsync
+@SpringBootApplication
 public class ManagercraftApplication {
 
     public static void main(String[] args) {
