+++ src/main/java/dev/gilbert/zacharia/managercraft/configuration/ShutdownListener.java
@@ -1,6 +1,6 @@
 package dev.gilbert.zacharia.managercraft.configuration;
 
-import dev.gilbert.zacharia.managercraft.secretsauce.minecraft.server.ServerManager;
+import dev.gilbert.zacharia.managercraft.secretsauce.minecraft.server.ServerProcessManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.event.ContextClosedEvent;
@@ -8,10 +8,10 @@ import org.springframework.context.event.ContextClosedEvent;
 public class ShutdownListener implements ApplicationListener<ContextClosedEvent> {
 
     @Autowired
-    private ServerManager serverManager;
+    private ServerProcessManager serverProcessManager;
 
     @Override
     public void onApplicationEvent(ContextClosedEvent event) {
-        serverManager.stopServer();
+        serverProcessManager.stopServer();
     }
 }
